+++ files/src/main/scala/dev/chronicles/files/versiontracker/FileSystemSyntax.scala
@@ -1,4 +1,4 @@
-package dev.chronicles.filebacked
+package dev.chronicles.files.versiontracker
 
 import java.nio.charset.StandardCharsets
