+++ files/src/main/scala/dev/chronicles/files/versiontracker/JsonCodecs.scala
@@ -1,11 +1,11 @@
-package dev.chronicles.filebacked
+package dev.chronicles.files.versiontracker
 
 import java.time.Instant
 
 import dev.chronicles.core.VersionTracker.TableOperation._
 import dev.chronicles.core.VersionTracker._
 import dev.chronicles.core.{Partition, TableName, Version}
-import dev.chronicles.filebacked.FileBackedVersionTracker.{StateFile, TableMetadataFile}
+import FileBackedVersionTracker.{StateFile, TableMetadataFile}
 import io.circe.syntax.EncoderOps
 import io.circe.{Decoder, Encoder, HCursor, Json}
