+++ files/src/test/scala/dev/chronicles/files/versiontracker/FileSystemSyntaxSpec.scala
@@ -1,4 +1,4 @@
-package dev.chronicles.filebacked
+package dev.chronicles.files.versiontracker
 
 import java.net.URI
 import java.nio.file.Files
